requirement.txt
altgraph==0.17
asgiref==3.2.7
attrs==19.3.0
beautifulsoup4==4.9.0
boto==2.49.0
boto3==1.13.3
botocore==1.16.3
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
-e git+https://github.com/django/django.git@6fbce45b0376f0ec8f3bf244f4f1f8d62c201f58#egg=Django
django-sphinx==2.2.4
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.14
ffpyplayer==4.2.0
futures==3.1.1
gensim==3.8.3
goslate==1.5.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.6.0
jmespath==0.9.5
Js2Py==0.70
Kivy==1.11.1
Kivy-Garden==0.1.4
macholib==1.14
more-itertools==8.2.0
numpy==1.16.1
opencv-python==4.2.0.34
packaging==20.3
pandas==0.24.1
patsy==0.5.1
pdfminer==20191125
pdfminer.six==20181108
pdfminer2==20151206
pdfminer3==2018.12.3.0
pdfminer3k==1.3.1
Pillow==7.1.2
pipwin==0.5.0
pluggy==0.13.1
ply==3.11
py==1.8.1
pycryptodome==3.9.7
Pygments==2.4.2
PyInstaller==3.6
pyjsparser==2.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
PyPrind==2.11.2
pySmartDL==1.3.3
pytesseract==0.3.4
pytest==5.4.2
python-dateutil==2.8.0
python-dotenv==0.13.0
pytz==2018.9
requests==2.22.0
s3transfer==0.3.3
scipy==1.4.1
six==1.12.0
smart-open==2.0.0
sortedcontainers==2.1.0
soupsieve==2.0
SpeechRecognition==3.8.1
sqlparse==0.3.0
statsmodels==0.9.0
tzlocal==2.1
urllib3==1.25.3
wcwidth==0.1.9
whitenoise==5.1.0
zipp==3.1.0
